Refactor profile registry to users domain
