+++ rpc/users/providers/services.py
@@ -14,7 +14,7 @@ from .models import (
   UsersProvidersCreateFromProvider1,
 )
 from server.modules.oauth_module import OauthModule
-from server.registry.accounts.profile import (
+from server.registry.users.profile import (
   get_profile_request,
   update_if_unedited_request,
 )
