+++ server/modules/session_module.py
@@ -8,7 +8,7 @@ from server.modules.auth_module import AuthModule, DEFAULT_SESSION_TOKEN_EXPIRY
 from server.modules.db_module import DbModule
 from server.modules.oauth_module import OauthModule
 from server.modules.discord_bot_module import DiscordBotModule
-from server.registry.accounts.profile import set_profile_image_request
+from server.registry.users.profile import set_profile_image_request
 from server.registry.security.accounts import get_security_profile_request
 from server.registry.security.sessions import (
   create_session_request,
