+++ server/registry/users/__init__.py
@@ -1,17 +1,22 @@
-"""Accounts domain registry bindings."""
+"""Users domain registry bindings."""
 
 from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+from . import profile
+
 if TYPE_CHECKING:
   from server.registry import RegistryRouter
 
 __all__ = [
+  "profile",
   "register",
 ]
 
 
 def register(router: "RegistryRouter") -> None:
-  """Register accounts domain routes."""
-  router.domain("accounts")
+  """Register users domain routes."""
+
+  domain = router.domain("users")
+  profile.register(domain.subdomain("profile"))
