+++ tests/test_auth_google_profile_image_update.py
@@ -6,7 +6,7 @@ from server.registry.types import DBRequest
 
 from server.modules.providers.auth.google_provider import GoogleAuthProvider
 from server.modules.oauth_module import OauthModule
-from server.registry.accounts.profile import set_profile_image_request
+from server.registry.users.profile import set_profile_image_request
 
 
 class DummyAuth:
