+++ tests/test_auth_microsoft_profile_refresh.py
@@ -5,7 +5,7 @@ from fastapi import FastAPI
 from server.registry.types import DBRequest
 
 from server.modules.oauth_module import OauthModule
-from server.registry.accounts.profile import update_if_unedited_request
+from server.registry.users.profile import update_if_unedited_request
 import json
